requirements.txt
Django==2.2.4
django-crispy-forms==1.8.0
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
sorl-thumbnail==12.7.0
django-allauth==0.43.0
django-grappelli==2.14.3
django-phone-field==1.8.1